Fix switching type not changing item dropdown and code cleanup.
For example, switching the type from 'uw' to 'artifact' did not update the dropdown for the item, and the item dropdown would be at the first alphabetical item. This was probably due to me not telling vue to update the dropdown values when the type changed. To fix this, I reverted back to using v-model with get and set.
Also removed some cruft that was no longer needed as well as cleaned up the vuex routing plugin.